% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signatures.R
\name{add_gene_signature_scores}
\alias{add_gene_signature_scores}
\title{Add Gene Signature Scores to Seurat Object}
\usage{
add_gene_signature_scores(seurat_obj, gene_set, Module_name = "Module_name")
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{gene_set}{A vector of genes of interest.}

\item{name}{A character object to describe the name of the gene_set.}
}
\value{
A Seurat object with added gene signature scores.
}
\description{
Add Gene Signature Scores to Seurat Object
}
