% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_gene_signature}
\alias{plot_gene_signature}
\title{Plot Gene Signature with plot_gene_signature}
\usage{
plot_gene_signature(
  seurat_obj,
  gene_set_name = "gene_set_score",
  reduction = "umap"
)
}
\arguments{
\item{seurat_obj}{A Seurat object with UMAP or TSNE computed.}

\item{gene_set_name}{Name of the  gene set score in the Seurat object metadata.}

\item{reduction}{Reduction pattern in the Seurat object.}
}
\value{
A UMAP or TSNE plot with coexpression of two gene signatures.
}
\description{
Plot Gene Signature with plot_gene_signature
}
