% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_two_gene_signature}
\alias{plot_two_gene_signature}
\title{Plot Gene Signature with plot_two_gene_signature}
\usage{
plot_two_gene_signature(
  seurat_obj,
  gene_set1_name = "gene_set1_score",
  gene_set2_name = "gene_set2_score",
  reduction = "umap"
)
}
\arguments{
\item{seurat_obj}{A Seurat object with UMAP or TSNE computed.}

\item{gene_set1_name}{Name of the first gene set score in the Seurat object metadata.}

\item{gene_set2_name}{Name of the second gene set score in the Seurat object metadata.}

\item{reduction}{Reduction pattern in the Seurat object.}
}
\value{
A UMAP or TSNE plot with coexpression of two gene signatures.
}
\description{
Plot Gene Signature with plot_two_gene_signature
}
